lane-backend/lane-api/requirements.txt
arrow==1.1.1
attrs==21.2.0
aws-lambda-builders==1.5.0
aws-wsgi==0.2.7
binaryornot==0.4.4
boto3==1.18.11
botocore==1.21.11
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
chevron==0.14.0
click==7.1.2
cookiecutter==1.7.3
cryptography==3.4.7
dateparser==0.7.6
docker==4.2.2
Flask==1.1.4
Flask-Cors==3.0.10
flask-lambda==0.0.4
freezegun==1.1.0
idna==2.10
importlib-metadata==4.6.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
more-itertools==8.8.0
moto==2.2.0
packaging==21.0
pluggy==0.13.1
poyo==0.5.0
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
pytest-mock==3.6.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
responses==0.13.3
s3transfer==0.5.0
serverlessrepo==0.1.10
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.7.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
watchdog==2.1.2
websocket-client==1.1.0
Werkzeug==1.0.1
xmltodict==0.12.0
zipp==3.5.0
